Move wagmi setup to separate file
