+++ src/utils/wagmi.ts
@@ -1,0 +1,53 @@
+import { configureChains, createClient, WagmiConfig } from 'wagmi'
+import { mainnet, polygon } from 'wagmi/chains'
+import { alchemyProvider } from 'wagmi/providers/alchemy'
+import { publicProvider } from 'wagmi/providers/public'
+
+import { connectorsForWallets } from '@rainbow-me/rainbowkit'
+import {
+  argentWallet,
+  braveWallet,
+  coinbaseWallet,
+  ledgerWallet,
+  metaMaskWallet,
+  rainbowWallet,
+  trustWallet,
+  walletConnectWallet,
+} from '@rainbow-me/rainbowkit/wallets'
+
+import { AlchemyApiKey } from 'constants/server'
+
+export const { chains, provider } = configureChains(
+  [mainnet, polygon],
+  [alchemyProvider({ apiKey: AlchemyApiKey }), publicProvider()]
+)
+
+const connectors = connectorsForWallets([
+  {
+    groupName: 'Recommended',
+    wallets: [
+      metaMaskWallet({ chains }),
+      rainbowWallet({ chains }),
+      argentWallet({ chains }),
+      coinbaseWallet({
+        appName: 'Index Coop',
+        chains,
+      }),
+      ledgerWallet({ chains }),
+    ],
+  },
+  {
+    groupName: 'Others',
+    wallets: [
+      walletConnectWallet({ chains }),
+      braveWallet({ chains }),
+      trustWallet({ chains }),
+    ],
+  },
+])
+
+export const wagmiClient = createClient({
+  autoConnect: true,
+  connectors,
+  provider,
+})
